# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: App greetings you!
      summary: Get greetings
      tags:
        - App
  /user:
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User created
        '400':
          description: Missing required fields
      summary: Create a user
      tags: &ref_0
        - User
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: List of users
      summary: Get all users
      tags: *ref_0
  /user/{id}:
    get:
      operationId: UserController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User
        '400':
          description: Invalid id
        '404':
          description: User not found
      summary: Get user by id
      tags: *ref_0
    put:
      operationId: UserController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: User updated
        '400':
          description: Invalid id
        '403':
          description: Previous password doesn't match
        '404':
          description: User not found
      summary: Update user by id
      tags: *ref_0
    delete:
      operationId: UserController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User deleted
        '400':
          description: Invalid id
        '404':
          description: User not found
      summary: Delete user by id
      tags: *ref_0
  /artist:
    post:
      operationId: ArtistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Artist created
        '400':
          description: Missing required fields
      summary: Create an artist
      tags: &ref_1
        - Artist
    get:
      operationId: ArtistController_findAll
      parameters: []
      responses:
        '200':
          description: List of artists
      summary: Get all artists
      tags: *ref_1
  /artist/{id}:
    get:
      operationId: ArtistController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Artist
        '400':
          description: Invalid id
        '404':
          description: Artist not found
      summary: Get artist by id
      tags: *ref_1
    put:
      operationId: ArtistController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Artist updated
        '400':
          description: Invalid id
        '404':
          description: Artist not found
      summary: Update artist by id
      tags: *ref_1
    delete:
      operationId: ArtistController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Artist deleted
        '400':
          description: Invalid id
        '404':
          description: Artist not found
      summary: Delete artist by id
      tags: *ref_1
  /track:
    post:
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: Track created
        '400':
          description: Missing required fields
      summary: Create a track
      tags: &ref_2
        - Track
    get:
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: List of tracks
      summary: Get all tracks
      tags: *ref_2
  /track/{id}:
    get:
      operationId: TrackController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Track
        '400':
          description: Invalid id
        '404':
          description: Track not found
      summary: Get track by id
      tags: *ref_2
    put:
      operationId: TrackController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: Track updated
        '400':
          description: Invalid id
        '404':
          description: Track not found
      summary: Update track by id
      tags: *ref_2
    delete:
      operationId: TrackController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Track deleted
        '400':
          description: Invalid id
        '404':
          description: Track not found
      summary: Delete track by id
      tags: *ref_2
  /album:
    post:
      operationId: AlbumController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Album created
        '400':
          description: Missing required fields
      summary: Create an album
      tags: &ref_3
        - Album
    get:
      operationId: AlbumController_findAll
      parameters: []
      responses:
        '200':
          description: List of albums
      summary: Get all albums
      tags: *ref_3
  /album/{id}:
    get:
      operationId: AlbumController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Album
        '400':
          description: Invalid id
        '404':
          description: Album not found
      summary: Get album by id
      tags: *ref_3
    put:
      operationId: AlbumController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Album updated
        '400':
          description: Invalid id
        '404':
          description: Album not found
      summary: Update album by id
      tags: *ref_3
    delete:
      operationId: AlbumController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Album deleted
        '400':
          description: Invalid id
        '404':
          description: Album not found
      summary: Delete album by id
      tags: *ref_3
  /favs/{entity}/{id}:
    post:
      operationId: FavoritesController_create
      parameters:
        - name: entity
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: '{entity} added to favorites'
      summary: Add an entity to favorites by id
      tags: &ref_4
        - Favorites
    delete:
      operationId: FavoritesController_remove
      parameters:
        - name: entity
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: '{entity} deleted from favorites'
      summary: Delete {entity} from favorites by id
      tags: *ref_4
  /favs:
    get:
      operationId: FavoritesController_findAll
      parameters: []
      responses:
        '200':
          description: List of favorites
      summary: Get all favorites
      tags: *ref_4
info:
  title: My API
  description: API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: SomeGreatLogin
          description: User login
        password:
          type: string
          example: '12345'
          description: User password
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: '12345'
          description: Previous user password
        newPassword:
          type: string
          example: new_password
          description: New user password
      required:
        - oldPassword
        - newPassword
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Michael Jackson
          description: Artist name
        grammy:
          type: boolean
          example: 'True'
          description: Shows if artist have ever won a Grammy Award
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Michael Jackson
          description: Artist name
        grammy:
          type: boolean
          example: 'True'
          description: Shows if artist have ever won a Grammy Award
      required:
        - name
        - grammy
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: The line
          description: Track name
        artistId:
          type: string
          example: 61f0c404-5cb3-11e7-907b-a6006ad3dba0
          description: Id of the artist to which the track belongs
        albumId:
          type: string
          example: 61f0c404-5cb3-11e7-907b-a6006ad3dba0
          description: Id of the album to which the track belongs
        duration:
          type: number
          example: '192'
          description: Track duration
      required:
        - name
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: The line
          description: Track name
        artistId:
          type: string
          example: 61f0c404-5cb3-11e7-907b-a6006ad3dba0
          description: Id of the artist to which the track belongs
        albumId:
          type: string
          example: 61f0c404-5cb3-11e7-907b-a6006ad3dba0
          description: Id of the album to which the track belongs
        duration:
          type: number
          example: '192'
          description: Track duration
      required:
        - name
        - duration
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Led Zeppelin III
          description: Album name
        year:
          type: string
          example: '1970'
          description: Album release date
        artistId:
          type: string
          example: 61f0c404-5cb3-11e7-907b-a6006ad3dba0
          description: Artist id
      required:
        - name
        - year
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Led Zeppelin III
          description: Album name
        year:
          type: string
          example: '1970'
          description: Album release date
        artistId:
          type: string
          example: 61f0c404-5cb3-11e7-907b-a6006ad3dba0
          description: Artist id
      required:
        - name
        - year